# api/docs/swagger.yaml
basePath: /api
definitions:
...
  docs.ErrorResponse:
    properties:
      error:
        example: Error message
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  docs.SuccessResponse:
    properties:
      data: {}
      success:
        example: true
        type: boolean
    type: object
  docs.TrashItem:
    properties:
      deletedAt:
        example: "2024-01-01T12:00:00Z"
        type: string
      deletedBy:
        example: admin
        type: string
      id:
        example: 1704067200000000000_document.pdf
        type: string
      isDir:
        example: false
        type: boolean
      name:
        example: document.pdf
        type: string
      originalPath:
        example: /home/admin/document.pdf
        type: string
      size:
        example: 1048576
        type: integer
    type: object
  docs.TrashListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/docs.TrashItem'
        type: array
      total:
        example: 5
        type: integer
      totalSize:
        example: 5242880
        type: integer
    type: object
  docs.TrashStatsResponse:
    properties:
      itemCount:
        example: 10
        type: integer
      newestItem:
        example: "2024-01-15T12:00:00Z"
        type: string
      oldestItem:
        example: "2024-01-01T12:00:00Z"
        type: string
      oldestItemDaysLeft:
        example: 15
        type: integer
      retentionDays:
        example: 30
        type: integer
      totalSize:
        example: 10485760
        type: integer
    type: object
...
host: localhost:3080
info:
  contact:
    email: support@simplecloudvault.io
    name: API Support
  description: |-
    Enterprise-grade cloud file sharing and management API.
    SimpleCloudVault provides comprehensive file management, sharing, and collaboration features.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: SimpleCloudVault API
  version: "1.0"
paths:
...
  /trash:
    get:
      consumes:
      - application/json
      description: Get all items in the user's trash, sorted by deletion time (newest
        first)
      produces:
      - application/json
      responses:
        "200":
          description: List of trash items
          schema:
            allOf:
            - $ref: '#/definitions/docs.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/docs.TrashListResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List trash items
      tags:
      - Trash
  /trash/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently delete an item from trash (irreversible)
      parameters:
      - description: Trash item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Item permanently deleted
          schema:
            $ref: '#/definitions/docs.SuccessResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "404":
          description: Trash item not found
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete from trash
      tags:
      - Trash
  /trash/{path}:
    post:
      consumes:
      - application/json
      description: Move a file or folder to the user's trash (soft delete)
      parameters:
      - description: Item path to move to trash
        in: path
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Item moved to trash
          schema:
            $ref: '#/definitions/docs.SuccessResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "404":
          description: Item not found
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Move to trash
      tags:
      - Trash
  /trash/empty:
    delete:
      consumes:
      - application/json
      description: Permanently delete all items from trash (irreversible)
      produces:
      - application/json
      responses:
        "200":
          description: Trash emptied
          schema:
            $ref: '#/definitions/docs.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Empty trash
      tags:
      - Trash
  /trash/restore/{id}:
    post:
      consumes:
      - application/json
      description: Restore an item from trash to its original location
      parameters:
      - description: Trash item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Item restored successfully
          schema:
            $ref: '#/definitions/docs.SuccessResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "404":
          description: Trash item not found
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Restore from trash
      tags:
      - Trash
  /trash/stats:
    get:
      consumes:
      - application/json
      description: Get statistics about the user's trash including item count, total
        size, and retention info
      produces:
      - application/json
      responses:
        "200":
          description: Trash statistics
          schema:
            allOf:
            - $ref: '#/definitions/docs.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/docs.TrashStatsResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get trash statistics
      tags:
      - Trash
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"